Split gallery CLI into fakes, fixtures, capture, and command modules
