+++ src/yoyopod/cli/pi/gallery/__init__.py
@@ -1,0 +1,5 @@
+"""Gallery CLI package for deterministic screen capture helpers."""
+
+from yoyopod.cli.pi.gallery.command import gallery_app
+
+__all__ = ["gallery_app"]
